Add options to tune figure size of gridsearch plots
As I used your package, I felt the need to be able to tune the figure size of the gridsearch plots.
I noticed that the cellwidth parameter was only available for plot_catstratpd_gridsearch. I added it to plot_stratpd_gridsearch.
I also added an extra parameter for the height to both gridsearch plot functions.